pyproject.toml
[project]
name = "peoplesoft-mcp-fin"
version = "0.1.0"
description = "PeopleSoft Finance MCP — query FSCM (GL, AP, AR, PO, AM) via semantic tools and SQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "fastmcp>=2.4.0",
    "mcp[cli]>=1.9.0",
    "oracledb>=3.1.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tools*"]

[tool.setuptools]
py-modules = ["peoplesoft_fin_server", "peoplesoft_server", "db", "llm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Dev deps (pytest): installed by default via [tool.uv] default-groups.
# Python download policy lives in uv.toml (no GitHub CPython fetch).
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
default-groups = ["dev"]
